split display.rs
