+++ src/symbols/binary_sym.rs
@@ -89,3 +89,11 @@ where
         BinarySym::new(v.0, v.1)
     }
 }
+
+impl<Op: BinaryOp, Sym1: DynamicSymbol<Out, In>, Sym2: DynamicSymbol<Out, In>, Out, In: ?Sized>
+    Display for BinarySym<Op, Sym1, Sym2, Out, In>
+{
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
+        Op::format_expression(f, |f| self.sym1.fmt(f), |f| self.sym2.fmt(f))
+    }
+}
